+++ CHANGELOG.md
@@ -9,8 +9,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- The server-side API `Connection::send_tor_exit_node_list` to send the list of
-  Tor exit nodes to the client.
+- The following server-side APIs to send data to the client:
+  - `Connection::send_blocklist`
+  - `Connection::send_tor_exit_node_list`
 
 ## [0.8.1] - 2024-11-15